move some types to a central file
